Role picker: split components into separate files
