+++ public/app/core/components/RolePicker/utils.ts
@@ -1,0 +1,5 @@
+import { Role } from 'app/types';
+
+export const isNotDelegatable = (role: Role) => {
+  return role.delegatable !== undefined && !role.delegatable;
+};
